eval : use uint16 label images

+++ pylene/src/contrib/segdet/evaluation.cpp
@@ -6,9 +6,9 @@
 
 namespace mln::contrib::segdet
 {
-  size_t image_max(mln::image2d<u_int8_t> img)
+  size_t image_max(mln::image2d<u_int16_t> img)
   {
-    uint8_t max = 0;
+    uint16_t max = 0;
 
     mln_foreach (auto v, img.values())
     {
@@ -19,9 +19,9 @@ namespace mln::contrib::segdet
     return max;
   }
 
-  mln::image2d<u_int8_t> image_where_equal(mln::image2d<u_int8_t> img, u_int8_t value, int& count)
+  mln::image2d<u_int16_t> image_where_equal(mln::image2d<u_int16_t> img, u_int16_t value, int& count)
   {
-    image2d<uint8_t> output = image2d<uint8_t>(img.width(), img.height());
+    image2d<uint16_t> output = image2d<uint16_t>(img.width(), img.height());
 
     mln_foreach (auto d, img.domain())
     {
@@ -37,13 +37,13 @@ namespace mln::contrib::segdet
     return output;
   }
 
-  mln::image2d<u_int8_t> image_where_equal(mln::image2d<u_int8_t> img, u_int8_t value)
+  mln::image2d<u_int16_t> image_where_equal(mln::image2d<u_int16_t> img, u_int16_t value)
   {
     int k = 0;
     return image_where_equal(img, value, k);
   }
 
-  int image_count_value(mln::image2d<u_int8_t> img, uint8_t value)
+  int image_count_value(mln::image2d<u_int16_t> img, uint16_t value)
   {
     int count = 0;
 
@@ -56,9 +56,9 @@ namespace mln::contrib::segdet
     return count;
   }
 
-  mln::image2d<u_int8_t> image_intersect(mln::image2d<uint8_t> img1, mln::image2d<uint8_t> img2)
+  mln::image2d<u_int16_t> image_intersect(mln::image2d<uint16_t> img1, mln::image2d<uint16_t> img2)
   {
-    image2d<uint8_t> out = image2d<uint8_t>(img1.width(), img1.height());
+    image2d<uint16_t> out = image2d<uint16_t>(img1.width(), img1.height());
 
     mln_foreach (auto d, img1.domain())
       out(d) = img1(d) * img2(d);
@@ -66,9 +66,9 @@ namespace mln::contrib::segdet
     return out;
   }
 
-  std::map<u_int8_t, int> image_unique(mln::image2d<u_int8_t> img)
+  std::map<u_int16_t, int> image_unique(mln::image2d<u_int16_t> img)
   {
-    std::map<u_int8_t, int> ret = {};
+    std::map<u_int16_t, int> ret = {};
 
     mln_foreach (auto v, img.values())
     {
@@ -87,8 +87,8 @@ namespace mln::contrib::segdet
 
   std::vector<double> score_detection(mln::ndbuffer_image ref_buf, mln::ndbuffer_image output_buf)
   {
-    mln::image2d<uint8_t> ref    = ref_buf.__cast<uint8_t, 2>();
-    mln::image2d<uint8_t> output = output_buf.__cast<uint8_t, 2>();
+    mln::image2d<uint16_t> ref    = ref_buf.__cast<uint16_t, 2>();
+    mln::image2d<uint16_t> output = output_buf.__cast<uint16_t, 2>();
 
     std::vector<double> seg_recall    = {};
     std::vector<double> seg_precision = {};
@@ -120,6 +120,7 @@ namespace mln::contrib::segdet
 
         auto labelled_only_segment = image_where_equal(output, kvp.first);
         auto seg_intersection      = image_intersect(labelled_only_segment, not_wanted_pixels);
+
         false_positive_count += width * height - image_count_value(seg_intersection, 0);
       }
